+++ core/library/src/commonMain/kotlin/ro/cosminmihu/ktor/monitor/ui/detail/DetailUiState.kt
@@ -92,6 +92,7 @@ internal data class DetailUiState(
 
     enum class FileShareType {
         Text,
+        Markdown,
     }
 
     enum class ContentFormat {