+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [TBA]
+### Added
+* Share as Markdown
+
 ## [1.16.0] - 2026-08-30
 ### Added
 * Bulk Share.